Update GCA2 website to reflect planning stage and remove logistics, etc. pages
The GCA2 website contained detained information about the meeting that
was to be held in 2020. This meeting was postponed and GCA2 is still in
planning stages. The text has been simplified to reflect this and additional
GCA2 pages have been deleted for now. They can be restored later via Git
if needed.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>